Moved DifferenceElement classes out of difference package to reduce visibility
