+++ javaparser-core/src/main/java/com/github/javaparser/printer/lexicalpreservation/Kept.java
@@ -1,4 +1,4 @@
-package com.github.javaparser.printer.lexicalpreservation.difference;
+package com.github.javaparser.printer.lexicalpreservation;
 
 import com.github.javaparser.TokenTypes;
 import com.github.javaparser.ast.type.PrimitiveType;
@@ -6,13 +6,11 @@ import com.github.javaparser.printer.concretesyntaxmodel.CsmElement;
 import com.github.javaparser.printer.concretesyntaxmodel.CsmIndent;
 import com.github.javaparser.printer.concretesyntaxmodel.CsmToken;
 import com.github.javaparser.printer.concretesyntaxmodel.CsmUnindent;
-import com.github.javaparser.printer.lexicalpreservation.DifferenceElementCalculator;
-import com.github.javaparser.printer.lexicalpreservation.LexicalDifferenceCalculator;
 
 public class Kept implements DifferenceElement {
     private final CsmElement element;
 
-    public Kept(CsmElement element) {
+    Kept(CsmElement element) {
         this.element = element;
     }
